requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-annotate --output-file=requirements.txt req.ing
#
aiohttp==3.7.4.post0
alabaster==0.7.12
appdirs==1.4.3
asciitree==0.3.3
async-timeout==3.0.1
atomicwrites==1.1.5
attrs==19.3.0
babel==2.8.0
backoff==1.10.0
black==19.10b0
cachetools==4.1.1
certifi==2020.6.20
cftime==1.2.1
chardet==3.0.4
click==7.1.2
cython==0.29.10
dacite==1.6.0
dask==2.10.1
dataclasses==0.6
decorator==4.4.2
docutils==0.16
entrypoints==0.3
f90nml==1.1.2
fasteners==0.15
flake8==3.7.8
fsspec==2021.11.0
fv3config==0.8.0
gcsfs==2021.11.0
google-api-core==1.22.1
google-auth-oauthlib==0.4.1
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-storage==1.23.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
h5netcdf==0.8.0
h5py==2.10.0
hypothesis==4.37.0
idna==2.8
imagesize==1.2.0
importlib-metadata==0.19
iniconfig==1.1.1
jinja2==2.11.3
joblib==1.0.1
kubernetes==10.0.1
markupsafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.2
multidict==5.1.0
netcdf4==1.4.2
numcodecs==0.7.2
numpy==1.16.2
oauthlib==3.1.0
pace-util==0.7.0
packaging==20.4
pandas==1.1.0
pathspec==0.8.0
pkgconfig==1.5.2
pluggy==0.12.0
protobuf==3.13.0
py==1.10.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pybind11==2.5.0
pycodestyle==2.5.0
pyflakes==2.1.1
pygments==2.6.1
pyparsing==2.4.7
pytest-regtest==1.4.5
pytest-subtests==0.4.0
pytest==6.2.3
python-dateutil==2.8.1
pytz==2020.1
pyyaml==5.1.2
regex==2020.7.14
requests-oauthlib==1.3.0
requests==2.22.0
rsa==4.6
scikit-learn==0.24.2
scipy==1.3.1
six==1.12.0
sklearn_json==0.1.0
snowballstemmer==2.0.0
sphinx-rtd-theme==0.5.0
sphinx==2.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
websocket-client==0.57.0
xarray==0.16.0
yarl==1.6.3
zarr==2.3.2
zipp==0.5.2

# The following packages are considered to be unsafe in a requirements file:
# setuptools
